fix: restaurant weekend API, migration 修正
